+++ tests/testthat/test-additive.R
@@ -1,5 +1,3 @@
-context("additive model")
-
 test_that("additive execution", {
   skip_on_cran()
 
@@ -44,7 +42,7 @@ test_that("additive execution", {
 
   # Extract the fit object from the trained workflow
   test_fit <- test_workflow_fit %>%
-    pull_workflow_fit()
+    extract_fit_parsnip()
 
   # Generate predictions for the data head from the trained workflow
   test_pred <- predict(test_workflow_fit, head(dat))
@@ -67,7 +65,7 @@ test_that("additive execution", {
   expect_equal(coef(test_fit$fit), coef(gam_fit))
 
   # Check the predictions
-  expect_equal(test_pred$.pred, gam_pred, tolerance = 0.1)
+  expect_equal(test_pred$.pred, gam_pred, ignore_attr = TRUE)
 
   # Check the default engine
   expect_equal(additive()$engine, "mgcv")