+++ NEWS.md
@@ -1,3 +1,11 @@
+# additive 0.0.3
+
+- Fixed `GetStarted.Rmd` vignette build
+- Fixed class predictions for binary classification
+- Fixed confidence intervals and standard errors
+- Added threshold probability option for class predictions
+- Replaced deprecated `pull_workflow_fit()`
+
 # additive 0.0.2
 
 - Added functionality to change to declare an engine in the model specification function.